refactor(commands): add init/learn/scaffold modules; thin CLI router; restore utils and generators; behavior unchanged; tests green
