+++ lib/generators/eslint-config.js
@@ -1,0 +1,49 @@
+"use strict";
+
+function generateEslintConfig() {
+  return `// Generated by eslint-plugin-ai-code-snifftest init
+import js from '@eslint/js';
+import aiSnifftest from 'eslint-plugin-ai-code-snifftest';
+
+export default [
+  js.configs.recommended,
+  {
+    files: ['**/*.js'],
+    plugins: { 'ai-code-snifftest': aiSnifftest },
+    rules: {
+      // Baseline
+      'no-unused-vars': ['warn', { argsIgnorePattern: '^_' }],
+      'no-undef': 'error',
+      'prefer-const': 'warn',
+      'no-var': 'error',
+      // Consistency
+      'quotes': ['warn', 'single', { avoidEscape: true }],
+      'semi': ['warn', 'always'],
+      'eqeqeq': ['error', 'always'],
+      // AI-friendly
+      'complexity': ['warn', 15],
+      'max-depth': ['warn', 4],
+      'max-lines-per-function': ['warn', 100],
+      // Naming (basic)
+      'camelcase': ['error', { properties: 'always' }],
+      // Domain-specific
+      'ai-code-snifftest/no-redundant-calculations': 'warn',
+      'ai-code-snifftest/no-equivalent-branches': 'warn',
+      'ai-code-snifftest/prefer-simpler-logic': 'warn',
+      'ai-code-snifftest/no-redundant-conditionals': 'warn',
+      'ai-code-snifftest/no-unnecessary-abstraction': 'warn',
+      'ai-code-snifftest/no-generic-names': 'warn',
+      'ai-code-snifftest/enforce-domain-terms': 'warn'
+    }
+  },
+  {
+    files: ['**/*.test.js', '**/*.spec.js'],
+    rules: { 'max-lines-per-function': 'off', 'complexity': 'off' }
+  }
+];
+`;
+}
+
+module.exports = {
+  generateEslintConfig
+};
